refactor(orgtrack-pm-cli): split command modules
Replace the monolithic commands.rs with a commands directory organized by command family and work-operation responsibility. Preserve the existing dispatch surface and wire behavior while keeping each implementation file focused and reviewable.

Verification: cargo fmt --check, cargo check --all-targets, cargo clippy --all-targets with warnings denied, and the full orgtrack-pm-cli test suite.

Pre-commit hook ran. Total eslint: 0, total circular: 0
